bringing check_logo manually
