+++ services_registry/utils/__init__.py
@@ -97,43 +97,3 @@ class Collector():
         except Exception as e:
             LOG.error("Invalid response for %s: %r", name, e)
             return (name, url, None, repr(e))
-
-def explore_service(name, url, info, error):
-    """Fetch the interesting information of a service
-    by using its base URL"""
-
-    # LOG.info("Exploring %s: %s", name, url)
-    # LOG.info("==> [error: %s] %s", error, info)
-
-    # import pprint
-    # pprint.pprint(info)
-
-    if error:
-        return {
-            "title": name,
-            "error": error,
-            "url": url
-        }
-
-    info = info.get('response',{}).get('results',{})
-    org = info.get("organization") or {}
-    beacon_id = info.get('id') or None
-    entities_json_file = f'static/entities/{beacon_id}.json';
-    d = {
-        "title": name,
-        "error": error,
-        "organization_name": org.get("name"),
-        "name": info.get("name"),
-        "description": info.get("description"),
-        "visit_us": org.get("welcomeUrl"),
-        "beacon_api": info.get("welcomeUrl"),
-        "contact_us": org.get("contactUrl"),
-        "logo_url": org.get("logoUrl", '')
-    }
-    try:
-        with open(entities_json_file) as fh:
-            entities = json.load(fh)
-            d["entities"] = entities[0]['entities']
-    except Exception as e:
-        LOG.error('Error on %s: %s', entities_json_file, e)
-    return d
